refactor: move aggregation back into file, improve code
